Apply heritrix-1.14.2 adjustments.
Apply adjustments to make commons-httpclient-3.0.1 work with heritrix-11.14.2
codebase. This involves using the classes found in the package
org.apache.commons.httpclient.heritrix.

Change the version to '3.0.1-heritrix-1.14.2-webcuratortool-2.0.1' in the
pom.xml, as this code is now based on commons-httpclient-3.0.1.

Updated the README.md to reflect the switch to 3.0.1.